Implement AuthorizationCodeCredential (formerly CodeFlowCredential)
